requirements.txt
async-timeout==4.0.2
blinker==1.6
cachelib==0.10.2
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
PyJWT==2.6.0
python-dotenv==1.0.0
redis==4.5.4
SQLAlchemy==2.0.8
typing_extensions==4.5.0
Werkzeug==2.2.3
